项目架构调整支持按位置划线模式
