+++ src/base_easy_marker.js
@@ -2,25 +2,20 @@ import Cursor, { CursorType } from './element/cursor'
 import Menu from './element/menu'
 import Mask from './element/mask'
 import Highlight from './element/highlight'
-import TextNode from './lib/text_node'
+
 import Markdown from './lib/markdown'
 import TouchEvent, { EventType } from './lib/touch_event'
 
 import {
-  getClickWordsPosition,
+  // getClickWordsPosition,
   getClickPosition,
   getTouchPosition,
   // getElementAbsolutePosition,
-  matchSubString,
-  screenRelativeToContainerRelative,
+  // matchSubString,
+  // screenRelativeToContainerRelative,
   anyToPx,
 } from './lib/helpers'
-
-const SelectStatus = {
-  NONE: 'none',
-  SELECTING: 'selecting',
-  FINISH: 'finish',
-}
+import { SelectStatus } from './lib/types'
 
 const defaultOptions = {
   menuItems: [],
@@ -178,7 +173,7 @@ class EasyMarker {
    * @param {Object} options.highlight highlight config
    * @param {string} options.highlight.color highlight color
    * @param {number} options.scrollSpeedLevel The speed of scrolling when touching bottom, default 4
-   * @param {number|string} options.scrollOffsetBottom The distance from the bottom when triggering scrolling，default 100
+   * @param {number|string} options.scrollOffsetBottom triggering scrolling, distance from the bottom, default 100
    * @param {Object} options.markdownOptions Customize options about the mapping relations between HTML and Markdown
    */
   constructor(options) {
@@ -196,19 +191,15 @@ class EasyMarker {
       start: null,
       end: null,
     }
-    this.textNode = {
-      start: null,
-      end: null,
-    }
+
     this.mask = null
     this.menu = null
-    this.markdown = null
     this.scrollOffsetBottom = null
     this.scrollSpeedLevel = null
+    this.containerScroll = null
     this.selectStatusChangeHandler = () => {}
     this.menuOnClick = () => {}
     this.highlightLineClick = () => {}
-    this.containerScroll = null
   }
 
   get selectStatus() {
@@ -353,29 +344,6 @@ class EasyMarker {
   }
 
   /**
-   * Get the selected text
-   *
-   * @memberof EasyMarker
-   * @returns {string}
-   */
-  getSelectText() {
-    const text =
-      TextNode.getSelectText(this.textNode.start, this.textNode.end) || ''
-    return matchSubString(this.container.innerText, text) || text
-  }
-
-  getSelectMarkdown() {
-    return (
-      this.markdown.getSelectMarkdown(
-        this.textNode.start.node,
-        this.textNode.end.node,
-        this.textNode.start.offset,
-        this.textNode.end.offset,
-      ).markdown || ''
-    )
-  }
-
-  /**
    * Highlight the lines between the specified nodes
    * @example
    * const id = 2;
@@ -514,6 +482,7 @@ class EasyMarker {
       start: null,
       end: null,
     }
+    // TODO base的destroy去触发各自的destroy
     this.textNode = {
       start: null,
       end: null,
@@ -526,6 +495,7 @@ class EasyMarker {
     this.selectStatus = SelectStatus.NONE
     this.cursor.start.hide()
     this.cursor.end.hide()
+    // TODO base的reset去触发各自的reset
     this.textNode = {
       start: null,
       end: null,
@@ -600,132 +570,6 @@ class EasyMarker {
   }
 
   /**
-   * Swap the start and end cursors
-   *
-   * @private
-   * @param {any} clickPosition
-   * @param {any} currentPosition
-   * @memberof EasyMarker
-   */
-  swapCursor(clickPosition, currentPosition) {
-    const { x, y } = currentPosition
-    if (this.movingCursor === this.cursor.start) {
-      const endPosition = this.cursor.end.position
-      if (y > endPosition.y || (y === endPosition.y && x >= endPosition.x)) {
-        this.cursor.start.position = this.cursor.end.position
-        this.movingCursor = this.cursor.end
-        this.textNode.start = new TextNode(
-          this.textNode.end.node,
-          this.textNode.end.offset,
-        )
-        this.textNode.end = new TextNode(
-          clickPosition.node,
-          clickPosition.index,
-        )
-      } else {
-        this.textNode.start = new TextNode(
-          clickPosition.node,
-          clickPosition.index,
-        )
-      }
-    } else {
-      const startPosition = this.cursor.start.position
-      if (
-        y < startPosition.y ||
-        (y === startPosition.y && x <= startPosition.x)
-      ) {
-        this.cursor.end.position = this.cursor.start.position
-        this.movingCursor = this.cursor.start
-        this.textNode.end = new TextNode(
-          this.textNode.start.node,
-          this.textNode.start.offset,
-        )
-        this.textNode.start = new TextNode(
-          clickPosition.node,
-          clickPosition.index,
-        )
-      } else {
-        this.textNode.end = new TextNode(
-          clickPosition.node,
-          clickPosition.index,
-        )
-      }
-    }
-  }
-
-  /**
-   * Start text selection
-   *
-   * @private
-   * @param {any} element
-   * @param {any} x
-   * @param {any} y
-   * @memberof EasyMarker
-   */
-  selectWords(element, x, y) {
-    const separators = [
-      '\u3002\u201D',
-      '\uFF1F\u201D',
-      '\uFF01\u201D',
-      '\u3002',
-      '\uFF1F',
-      '\uFF01',
-    ]
-    const {
-      rects, node, index, wordsLength,
-    } =
-      getClickWordsPosition(element, x, y, separators) || {}
-    if (!rects || (rects && rects.length === 0)) return
-
-    const startRect = rects[0]
-    const endRect = rects[rects.length - 1]
-    // start
-    const startLeft = startRect.left - this.screenRelativeOffset.x
-    const startTop = startRect.top - this.screenRelativeOffset.y
-    this.textNode.start = new TextNode(node, index)
-    this.cursor.start.height = startRect.height
-    this.cursor.start.position = { x: startLeft, y: startTop }
-
-    // end
-    const endLeft = endRect.left - this.screenRelativeOffset.x
-    const endTop = endRect.top - this.screenRelativeOffset.y
-    this.textNode.end = new TextNode(node, index + wordsLength)
-    this.cursor.end.height = endRect.height
-    this.cursor.end.position = { x: endLeft + endRect.width, y: endTop }
-
-    this.cursor.start.show()
-    this.cursor.end.show()
-
-    this.renderMask()
-    this.selectStatus = SelectStatus.FINISH
-  }
-
-  /**
-   * Renders the selected mask layer
-   * @private
-   * @memberof EasyMarker
-   */
-  renderMask() {
-    const { header, body, footer } = TextNode.getSelectRects(
-      this.textNode.start,
-      this.textNode.end,
-    )
-    const relativeHeader = screenRelativeToContainerRelative(
-      header,
-      this.screenRelativeOffset,
-    )
-    const relativeBody = screenRelativeToContainerRelative(
-      body,
-      this.screenRelativeOffset,
-    )
-    const relativeFooter = screenRelativeToContainerRelative(
-      footer,
-      this.screenRelativeOffset,
-    )
-    this.mask.render(relativeHeader, relativeBody, relativeFooter)
-  }
-
-  /**
    *
    * @private
    * @param {HTMLElement} element
@@ -736,39 +580,6 @@ class EasyMarker {
       this.container.contains(element) &&
       this.excludeElements.findIndex(el => el.contains(element)) === -1
     )
-  }
-
-  /**
-   * Tap event
-   *
-   * @private
-   * @param {TouchEvent} e
-   * @memberof EasyMarker
-   */
-  handleTap(e) {
-    if (this.selectStatus === SelectStatus.FINISH) {
-      this.menu.handleTap(e, {
-        start: this.textNode.start,
-        end: this.textNode.end,
-        content: this.getSelectText(),
-        markdown: this.getSelectMarkdown(),
-      })
-      const position = this.getTouchRelativePosition(e)
-      const startCursorRegion = this.cursor.start.inRegion(position)
-      const endCursorRegion = this.cursor.end.inRegion(position)
-      if (startCursorRegion.inRegion || endCursorRegion.inRegion) return
-      this.reset()
-    } else if (this.selectStatus === SelectStatus.NONE) {
-      const inHighlightLine = this.highlight.handleTap(e)
-      if (
-        !inHighlightLine &&
-        !this.options.disableTapHighlight &&
-        this.isContains(e.target)
-      ) {
-        const { x, y } = getTouchPosition(e)
-        this.selectWords(e.target, x, y)
-      }
-    }
   }
 
   /**
