+++ src/index.js
@@ -1,0 +1,3 @@
+import NodeEasyMarker from './node_easy_marker'
+
+export default NodeEasyMarker
