+++ src/lib/types.js
@@ -1,0 +1,12 @@
+
+const SelectStatus = {
+  NONE: 'none',
+  SELECTING: 'selecting',
+  FINISH: 'finish',
+}
+
+const EasyMarkerMode = {
+  NODE: 'node',
+  REGION: 'region',
+}
+export { SelectStatus, EasyMarkerMode }
